refactor: update server and swagger configurations
- Renamed server and swagger configuration files for consistency.
- Removed deprecated server and swagger files to streamline the codebase.
- Adjusted import paths in the configuration barrel to reflect the new file names.
