+++ src/config/index.js
@@ -1,3 +1,3 @@
 // Config barrel exports
-export { createServer } from './server.js';
-export { swaggerConfig, swaggerUIConfig } from './swagger.js';
+export { createServer } from './server.config.js';
+export { swaggerConfig, swaggerUIConfig } from './swagger.config.js';
